requirements.txt
Django==4.1.1
Pillow==9.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.3
django-crispy-forms==1.14.0
requests==2.28.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-redis==5.2.0
gunicorn==21.2.0
whitenoise==6.6.0
